Clean up type signatures of syscalls.
Instead of returning long values, prefer int.

If a function returns a byte count or an error, return ssize_t.

Make handle_t signed so that it can be used as a return value with
possible error.

If an argument is a handle, make it a handle_t rather than uint32_t.

This makes the signature of close(...) compatible with POSIX.

Bug: 138738352

+++ lib/libc-trusty/include/sys/mman.h
@@ -28,11 +28,11 @@ extern "C" {
 #endif
 
 /* Trusty specific. */
-long prepare_dma(void* uaddr,
-                 uint32_t size,
-                 uint32_t flags,
-                 struct dma_pmem* pmem);
-long finish_dma(void* uaddr, uint32_t size, uint32_t flags);
+int prepare_dma(void* uaddr,
+                uint32_t size,
+                uint32_t flags,
+                struct dma_pmem* pmem);
+int finish_dma(void* uaddr, uint32_t size, uint32_t flags);
 
 #ifdef __cplusplus
 }
